[Design] #17 온보딩 기능 데이터 연결
