+++ Team12Snail/Team12Snail/Source/View/ContentView.swift
@@ -9,9 +9,9 @@ import SwiftUI
 
 struct ContentView: View {
     var body: some View {
-        //PlanetView()
+        PlanetView()
         //StopWatchView()
-        OnboardingView()
+//        OnboardingView()
     }
 }
