+++ Team12Snail/Team12Snail/Source/View/Planet/PlanetView.swift
@@ -32,12 +32,6 @@ struct PlanetView: View {
             ShoppingButton
                 .padding()
         }
-        .onAppear {
-            // 현재는 임시로 뷰 생성시 Items 임의로 생성되게 설정
-            if items.isEmpty {
-                initItems()
-            }
-        }
     }
 }
 
@@ -102,28 +96,4 @@ extension PlanetView {
 
 #Preview {
     PlanetView()
-}
-
-extension PlanetView {
-    
-    // 원래는 온보딩 뷰에서 불러와야하는 메소드
-    func initItems() {
-        let item = Item(
-            starName: [0:"첫번째", 1:"두번째", 2:"세번째", 3:"네번째"],
-            starPoint: [0:0, 1:0, 2:0, 3:0],
-            selectedItems: [0 : "pencil", 1 : "eraser", 2 : "paperplane.fill", 3 : "doc.fill"],
-            purchasedItems: [
-                0 : [],
-                1 : [],
-                2 : [],
-                3 : []
-            ])
-        modelContext.insert(item)
-        
-        do {
-            try modelContext.save()
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
 }
